chore: 기능별 파일 분리
열거형, 클래스, 선언부, 코드실행부로 각각 코드 분리
