+++ RockPaperScissors/RockPaperScissors/Prepare/prepareGame.swift
@@ -1,0 +1,11 @@
+//
+//  gamesReady.swift
+//  RockPaperScissors
+//
+//  Created by 박세웅 on 2022/02/17.
+//
+
+import Foundation
+
+var rockPaperScissorsGame: RockPaperScissors = RockPaperScissors(user: user, computer: computer)
+var mukChiBaGame: MukChiBa = MukChiBa(user: user, computer: computer)
