+++ RockPaperScissors/RockPaperScissors/Prepare/preparePlayer.swift
@@ -1,0 +1,11 @@
+//
+//  playerReady.swift
+//  RockPaperScissors
+//
+//  Created by 박세웅 on 2022/02/17.
+//
+
+import Foundation
+
+var user: Player = Player(name: "사용자")
+var computer: Player = Player(name: "컴퓨터")
